changed and shifted unit tests to suggested files
