+++ tests/unit/test_models/test_full_battery_models/test_lithium_ion/base_lithium_ion_tests.py
@@ -2,6 +2,7 @@
 # Base unit tests for the lithium-ion models
 #
 import pybamm
+import pytest
 
 
 class BaseUnitTestLithiumIon:
@@ -54,6 +55,18 @@ class BaseUnitTestLithiumIon:
             "thermal": "lumped",
         }
         self.check_well_posedness(options)
+
+    def test_well_posed_lumped_thermal_capacity_model(self):
+        options = {"thermal": "lumped", "use lumped thermal capacity": "true"}
+        self.check_well_posedness(options)
+
+    def test_incompatible_lumped_thermal_capacity_option(self):
+        options = {"thermal": "x-full", "use lumped thermal capacity": "true"}
+        with pytest.raises(
+            pybamm.OptionError,
+            match="Lumped thermal capacity model only compatible with lumped thermal models",
+        ):
+            self.check_well_posedness(options)
 
     def test_well_posed_thermal_1plus1D(self):
         options = {
